update: validation

+++ dist /controllers/user.controller.js
@@ -1,6 +1,10 @@
 "use strict";
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.userController = void 0;
+const mongoose_1 = __importDefault(require("mongoose"));
 const errors_1 = require("../errors");
 const services_1 = require("../services");
 const validations_1 = require("../validations");
@@ -16,7 +20,7 @@ class UserController {
     }
     async create(req, res, next) {
         try {
-            const { error, value } = validations_1.userSchema.validate(req.body);
+            const { error, value } = validations_1.userSchema.create.validate(req.body);
             if (error) {
                 throw new errors_1.ApiError("Validation failed", 400);
             }
@@ -32,6 +36,9 @@ class UserController {
     async findById(req, res, next) {
         try {
             const { id } = req.params;
+            if (!mongoose_1.default.isObjectIdOrHexString(id)) {
+                throw new errors_1.ApiError("Not valid Id", 400);
+            }
             const user = await services_1.userService.findById(id);
             return res.status(200).json({ data: user });
         }
@@ -42,7 +49,10 @@ class UserController {
     async updateByIdPut(req, res, next) {
         try {
             const { id } = req.params;
-            const { value } = validations_1.userSchema.validate(req.body);
+            if (!mongoose_1.default.isObjectIdOrHexString(id)) {
+                throw new errors_1.ApiError("Not valid Id", 400);
+            }
+            const { value } = validations_1.userSchema.create.validate(req.body);
             const updatedUser = await services_1.userService.updateByIdPut(id, value);
             return res
                 .status(200)
@@ -55,7 +65,10 @@ class UserController {
     async updateByIdPatch(req, res, next) {
         try {
             const { id } = req.params;
-            const { value } = validations_1.updateUserSchema.validate(req.body);
+            if (!mongoose_1.default.isObjectIdOrHexString(id)) {
+                throw new errors_1.ApiError("Not valid Id", 400);
+            }
+            const { value } = validations_1.userSchema.updateUserSchema.validate(req.body);
             const updatedUser = await services_1.userService.updateByIdPatch(id, value);
             return res
                 .status(200)
@@ -68,6 +81,9 @@ class UserController {
     async deleteById(req, res, next) {
         try {
             const { id } = req.params;
+            if (!mongoose_1.default.isObjectIdOrHexString(id)) {
+                throw new errors_1.ApiError("Not valid Id", 400);
+            }
             await services_1.userService.deleteById(id);
             return res.status(200).json({ message: `User id=${id} is deleted` });
         }
